requirements.txt
# Core dependencies
pyyaml==6.0.1

# Python version requirement
python>=3.10